Going back to using int movie id as primary key causes the endless loading of page 2 again as opposed to Long auto generated one

+++ app/src/main/java/com/asmat/rolando/nightwing/popular_movies/PopularMoviesRemoteMediator.kt
@@ -22,7 +22,7 @@ class PopularMoviesRemoteMediator(
     val moviesDao = database.moviesDAO()
 
     // The page that a movie ID belongs to
-    val keysTable = mutableMapOf<Long, RemoteKeys>()
+    val keysTable = mutableMapOf<Int, RemoteKeys>()
 
     class RemoteKeys(val previousKey: Int?, val nextKey: Int?)
